move DVID label/supervoxel access to client
