+++ dvid_point_cloud/__init__.py
@@ -3,5 +3,5 @@
 __version__ = "0.1.0"
 
 from .client import DVIDClient
-from .sampling import uniform_auto_scale, uniform_sample, sample_for_bodies
+from .sampling import uniform_auto_scale, uniform_sample
 from .neuroglancer import point_cloud_to_neuroglancer_json
